+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { default as Loading } from './Loading';
 export { default as SortableList } from './SortableList';
+export { default as ReferredRecord } from './ReferredRecord/ReferredRecord';